mcar_test upload
The mcar_test.  It basically splits the dataset into two components
based upon whether a given variables is present or absence, and then
compares the expected mean (if continuous) or count (if categorical).

It still has a few bugs (doesn’t work on some datasets), and is quite
slow.

I’m slowly adding in the roxygen documentation.