import packages
I forgot to include this into the description file, and subsequently
when I went to import packages using `@import dplyr`, and so the
package wouldn’t load those packages. I found this thanks to the
roxygen2 documentation -
http://cran.r-project.org/web/packages/roxygen2/vignettes/namespace.html

I’m not sure if there is a limit to the number of imported packages to
include, or what the ‘Suggests:’ command does in Description.